Minor fixes, optimizing
